refactor(server): extract hook adapters from main.rs into hooks module (7792 -> 7375 lines)
Audit 2026-05-19 P3 partial. main.rs's 6 delivery-bus hook adapters
(KmsHookAdapter, SesEmailDispatcher, SesSendEmailDispatcherImpl,
Elbv2TargetRegistrationImpl, EcsTaskRunnerImpl, SnsSmsDispatcher)
plus their trait impls were sitting at the tail of main.rs at lines
7153-7690 — moved them to a sibling `hooks` module.

Remaining 7375 main.rs lines are still dominated by the giant
async fn main() boot fn (~7050 lines); breaking that down requires
actual refactoring rather than mechanical extraction.
